extract platform selector into single file
