+++ vllm/spec_decode/multi_step_worker.py
@@ -15,25 +15,11 @@ if current_platform.is_cuda_alike():
 from vllm.spec_decode.interfaces import (SpeculativeProposals,
                                          SpeculativeProposer)
 from vllm.spec_decode.proposer_worker_base import ProposerWorkerBase
+from vllm.spec_decode.selector import WorkerCls
 from vllm.spec_decode.top1_proposer import Top1Proposer
 
-if current_platform.is_neuron():
-    from vllm.worker.neuron_worker import NeuronWorker as WorkerBaseCls
-elif current_platform.is_hpu():
-    from vllm.worker.hpu_worker import HPUWorker as WorkerBaseCls
-elif current_platform.is_openvino():
-    from vllm.worker.openvino_worker import OpenVINOWorker as WorkerBaseCls
-elif current_platform.is_cpu():
-    from vllm.worker.cpu_worker import CPUWorker as WorkerBaseCls
-elif current_platform.is_tpu():
-    from vllm.worker.tpu_worker import TPUWorker as WorkerBaseCls
-elif current_platform.is_xpu():
-    from vllm.worker.xpu_worker import XPUWorker as WorkerBaseCls
-else:
-    from vllm.worker.worker import Worker as WorkerBaseCls
 
-
-class MultiStepWorker(WorkerBaseCls, ProposerWorkerBase):
+class MultiStepWorker(WorkerCls, ProposerWorkerBase):
     """The MultiStepWorker is equivalent to a Worker except that it allows
     multiple forward passes in a single call, assuming the scheduler has
     allocated enough space to store the additional KV. This reduces overhead
