+++ vllm/spec_decode/target_model_runner.py
@@ -1,39 +1,8 @@
 from typing import List, Optional
 
 from vllm.config import VllmConfig
-from vllm.platforms import current_platform
 from vllm.sequence import SequenceGroupMetadata
-
-if current_platform.is_cuda_alike():
-    from vllm.worker.model_runner import (
-        ModelInputForGPUWithSamplingMetadata as ModelInputCls)  # yapf: disable
-    from vllm.worker.model_runner import ModelRunner as ModelRunnerCls
-elif current_platform.is_neuron():
-    from vllm.worker.neuron_model_runner import (
-        ModelInputForNeuron as ModelInputCls)  # yapf: disable
-    from vllm.worker.neuron_model_runner import (
-        NeuronModelRunner as ModelRunnerCls)  # yapf: disable
-elif current_platform.is_hpu():
-    from vllm.worker.hpu_model_runner import HPUModelRunner as ModelRunnerCls
-    from vllm.worker.hpu_model_runner import (
-        ModelInputForHPUWithSamplingMetadata as ModelInputCls)  # yapf: disable
-elif current_platform.is_openvino():
-    from vllm.worker.openvino_model_runner import ModelInput as ModelInputCls
-    from vllm.worker.openvino_model_runner import (
-        OpenVINOModelRunner as ModelRunnerCls)  # yapf: disable
-elif current_platform.is_cpu():
-    from vllm.worker.cpu_model_runner import CPUModelRunner as ModelRunnerCls
-    from vllm.worker.cpu_model_runner import (
-        ModelInputForCPUWithSamplingMetadata as ModelInputCls)  # yapf: disable
-elif current_platform.is_tpu():
-    from vllm.worker.tpu_model_runner import ModelInputForTPU as ModelInputCls
-    from vllm.worker.tpu_model_runner import TPUModelRunner as ModelRunnerCls
-elif current_platform.is_xpu():
-    from vllm.worker.xpu_model_runner import (
-        ModelInputForXPUWithSamplingMetadata as ModelInputCls)  # yapf: disable
-    from vllm.worker.xpu_model_runner import XPUModelRunner as ModelRunnerCls
-else:
-    raise ValueError(f"Unsupported platform: {current_platform}")
+from vllm.spec_decode.selector import ModelInputCls, ModelRunnerCls
 
 
 class TargetModelRunner(ModelRunnerCls):
